apps/console-test/pyproject.toml
[tool.poetry]
name = "console-test"
version = "0.1.0"
description = ""
authors = ["Matthew Russell <mattrussell36@gmail.com>"]
readme = "README.md"
packages = [{include = "console_test"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
psutil = "^5.9.5"
vega-sim = {git = "https://github.com/vegaprotocol/vega-market-sim.git"}
pytest-playwright = "^0.4.2"
docker = "^6.1.3"
pytest-xdist = "^3.3.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
timeout = 120
log_cli = true
log_cli_format = "%(asctime)s - %(name)s - %(levelname)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_level = "INFO"
log_file_format = "%(asctime)s - %(name)s - %(levelname)s: %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
log_file_level = "INFO"